refactor: split fileUtils.ts (772 LOC) and config.ts (648 LOC) into focused sub-modules (#51)
Resolves #51. Both files exceeded the 300 LOC limit by 2x+. Each has been
decomposed into sub-modules under src/utils/files/ and src/utils/config/,
with the original files reduced to thin backward-compatible re-export facades.

fileUtils.ts (772 → 142 LOC):
- src/utils/files/types.ts         - FileOperationError, FileMetadata, CleanupOptions, ArchiveOptions, SyncOptions
- src/utils/files/FileReader.ts    - exists, readJsonFile, getMetadata, getDirectorySize
- src/utils/files/FileWriter.ts    - ensureDirectory, writeJsonFile, appendToFile, move, delete, createTemp*
- src/utils/files/FileSearch.ts    - findFiles, calculateHash, filterByAge, filterByPatterns
- src/utils/files/FileArchiver.ts  - copy, backup, cleanup, sync
- Removed 7 duplicate standalone functions (ensureDir, readJson, writeJson, exists, remove, copy, findFiles)
  that wrapped class static methods without adding value

config.ts (648 → 97 LOC):
- src/utils/config/types.ts          - ConfigError, ConfigSource, ValidationResult, ConfigMetadata, DEFAULT_CONFIG, ENV_MAPPINGS
- src/utils/config/ConfigValidator.ts - validateConfig() pure validation function
- src/utils/config/ConfigLoader.ts   - loadConfigFile, loadEnvConfig, mergeConfigs, getNestedValue, setNestedValue
- src/utils/config/ConfigManager.ts  - ConfigManager class with get/set/watch/load/export
- All factory functions preserved in facade (createConfigManager, loadDefaultConfig, getConfig, updateConfig)

All 36 config tests pass. tsc --noEmit passes with zero errors.
