+++ src/utils/config/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Config sub-modules barrel export
+ */
+
+export * from './types';
+export * from './ConfigValidator';
+export * from './ConfigLoader';
+export * from './ConfigManager';
