+++ src/utils/files/index.ts
@@ -1,0 +1,9 @@
+/**
+ * File utilities sub-modules barrel export
+ */
+
+export * from './types';
+export * from './FileReader';
+export * from './FileWriter';
+export * from './FileSearch';
+export * from './FileArchiver';
